Create unit tests

+++ src/main/java/com/n26/domain/Transaction.java
@@ -3,20 +3,21 @@ package com.n26.domain;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Transaction implements Serializable {
-    private final double amount;
+    private final BigDecimal amount;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSXXX")
     private final Date timestamp;
 
-    public Transaction(double amount, Date timestamp) {
+    public Transaction(BigDecimal amount, Date timestamp) {
         this.amount = amount;
         this.timestamp = timestamp;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
